Add files via upload
This Jupyter notebook, authored by Trey Martin on October 22, 2024, focuses on analyzing health spending data as a percentage of GDP. The project processes a raw dataset, filtering it to identify the top 10 countries with the highest health spending for the years 2016–2021. Using Python libraries like Pandas and Matplotlib, the notebook performs data manipulation to create organized dataframes and generate a graph visualizing the findings. The visualization highlights trends and insights in health expenditure among these leading countries, and the final graph is saved as a PNG file for further use in reporting or presentations.